Moved the tests to test.py
